insane: Add QA check for RRECOMMENDS on non-existent packages
Add a package QA check for when a package RRECOMMENDS another that won't
be built because it is empty and ALLOW_EMPTY is not set. This happens
usually when ${PN}-dev RRECOMMENDS ${PN} but ${PN} is empty. This is not
an error but might be something to look into.

Example of a generated warning:
WARNING: python3-3.11.2-r0 do_package_qa:
QA Issue: python3-dev -> python3 but won't be build [rrecommends-non-existent]

Note: This QA check is voluntarily not enabled by default.

This will make bugs [YOCTO #8222] and [YOCTO #6839] appear in the QA logs.

Co-authored-by: Fawzi Khaber <fawzi.khaber@smile.fr>
(From OE-Core rev: 779870750e00be77c921436a0d860004c477b14f)

Signed-off-by: Yoann Congal <yoann.congal@smile.fr>
Signed-off-by: Alexandre Belloni <alexandre.belloni@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>